requirements.txt
streamlit>=1.36
openai>=1.50.2
pypdf>=4.2.0
python-dotenv>=1.0.1
